+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Option `sasl_username`
 - Event 901 `RPL_LOGGEDOUT`
 - Event 902 `ERR_NICKLOCKED`
+- Event 903 `RPL_SASLSUCCESS`
 - Event 904 `ERR_SASLFAIL`
 - Event 905 `ERR_SASLTOOLONG`
 - Event 907 `ERR_SASLALREADY`